[ 기본 과제 ] : 스토리보드 분리하기 #4
##### 제목은 최대 50 글자까지만 입력 ############## -> |
# 본문은 위에 작성
######## 본문은 한 줄에 최대 72 글자까지만 입력 ########################### -> |
# 꼬릿말은 아래에 작성: ex) #이슈 번호
# --- COMMIT END ---
#- ✅[CHORE] : 코드 수정, 내부 파일 수정
#- ✨[FEAT] : 새로운 기능 구현
#- ➕[ADD] : Feat 이외의 부수적인 코드 추가, 라이브러리 추가, 새로운 파일 생성 시
#- 🍱[ADD] : 에셋 추가
#- 🚑️[HOTFIX] : issue나, QA에서 급한 버그 수정에 사용
#- 🔨[FIX] : 버그, 오류 해결
#- ⚰️ [DEL] : 쓸모없는 코드 삭제
#- 📝[DOCS] : README나 WIKI 등의 문서 개정
#- 💄[MOD] : storyboard 파일,UI 수정한 경우
#- ✏️ [CORRECT] : 주로 문법의 오류나 타입의 변경, 이름 변경 등에 사용합니다.
#- 🚚[MOVE] : 프로젝트 내 파일이나 코드의 이동
#- ⏪️[RENAME] : 파일 이름 변경이 있을 때 사용합니다.
#- ⚡️[IMPROVE] : 향상이 있을 때 사용합니다.
#- ♻️[[REFACTOR] : 전면 수정이 있을 때 사용합니다
#- 🔀[MERGE]: 다른브렌치를 merge 할 때 사용합니다.
# ------------------
#     제목 첫 글자를 대문자로
#     제목은 명령문으로
#     제목 끝에 마침표(.) 금지
#     제목과 본문을 한 줄 띄워 분리하기
#     본문은 "어떻게" 보다 "무엇을", "왜"를 설명한다.
#     본문에 여러줄의 메시지를 작성할 땐 "-"로 구분
# ------------------
