+++ instagramLogin/instagramLogin/Password/PasswordViewController.swift
@@ -1,5 +1,5 @@
 //
-//  UsernameViewController.swift
+//  PasswordViewController.swift
 //  instagramLogin
 //
 //  Created by 황찬미 on 2022/04/04.
@@ -7,16 +7,18 @@
 
 import UIKit
 
-class UsernameViewController: UIViewController {
-
-    @IBOutlet weak var nameTextField: UITextField!
+class PasswordViewController: UIViewController {
+    
+    var nameData: String?
+    
     @IBOutlet weak var nextButton: UIButton!
+    @IBOutlet weak var pwTextField: UITextField!
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         setButton()
         changeBackButton()
+        
     }
     
     private func changeBackButton() {
@@ -28,27 +30,31 @@ class UsernameViewController: UIViewController {
       self.navigationController?.popViewController(animated: true)
     }
     
-    
-    // 다음 버튼, textfield의 데이터를 다음 VC로 넘겨 준다. push
+    // 다음 버튼, 입력받은 데이터를 다음 VC를 modal로 present 함.
     @IBAction func nextButton(_ sender: Any) {
-        guard let nextVC = self.storyboard?.instantiateViewController(withIdentifier: "PasswordViewController") as? PasswordViewController else { return }
         
-        nextVC.nameData = nameTextField.text
+        let completionSB = UIStoryboard(name: "Completion", bundle: nil)
+        let completionVC = completionSB.instantiateViewController(withIdentifier: "CompletionViewController") as! CompletionViewController
+    
+        // 전달받은 데이터의 값이 있으면, 다음 VC 프로퍼티로 값을 넘겨 준다
+        if let nameData = nameData {
+            completionVC.nameData = nameData
+        }
         
-        self.navigationController?.pushViewController(nextVC, animated: true)
+        self.present(completionVC, animated: true, completion: nil)
     }
+    
+    
 
-
-    @IBAction func nameEdit(_ sender: Any) {
+    @IBAction func passEdit(_ sender: Any) {
         setButton()
     }
     
     private func setButton() {
-        if nameTextField.hasText {
+        if pwTextField.hasText {
             nextButton.isEnabled = true
         } else {
             nextButton.isEnabled = false
         }
     }
-    
 }
