+++ instagramLogin/instagramLogin/Welcome/WelcomeViewController.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class WelcomeViewController: UIViewController {
     
     var nameData: String?
     let button = UIButton(type: .custom)
@@ -31,8 +31,8 @@ class ViewController: UIViewController {
     
     // 가입하기 버튼, 다음 VC로 데이터 push
     @IBAction func joinButton(_ sender: Any) {
-        guard let userNameVC = self.storyboard?.instantiateViewController(withIdentifier: "UsernameViewController") as? UsernameViewController else { return }
-        
+        let usernameSB = UIStoryboard(name: "Username", bundle: nil)
+        let userNameVC = usernameSB.instantiateViewController(withIdentifier: "UsernameViewController") as! UsernameViewController
         
         self.navigationController?.pushViewController(userNameVC, animated: true)
     }
@@ -40,11 +40,10 @@ class ViewController: UIViewController {
     
     // 로그인 버튼, 마지막 완료 화면으로 modal로 present
     @IBAction func loginButton(_ sender: Any) {
-        guard let joinVC = self.storyboard?.instantiateViewController(withIdentifier: "CompletionViewController") as? CompletionViewController else { return }
+        let completionSB = UIStoryboard(name: "Completion", bundle: nil)
+        let completionVC = completionSB.instantiateViewController(withIdentifier: "CompletionViewController") as! CompletionViewController
         
-        joinVC.nameData = nameTextfield.text
-        
-        self.present(joinVC, animated: true, completion: nil)
+        self.present(completionVC, animated: true, completion: nil)
     }
     
     @IBAction func nameEdit(_ sender: Any) {
